+++ Simulation.cpp
@@ -37,7 +37,7 @@ Simulation::Simulation()
 
 Simulation::~Simulation()
 {
-	for (int i = 0; i < players.size(); i++)
+	for (size_t i = 0; i < players.size(); i++)
 	{
         delete players[i];
 	}
